feat: implement 10 tutorial-verified gap tracks in parallel
Ten independent API tracks addressing gaps surfaced by the deep tutorial
analysis + adversarial verification. 3710 insertions, 285 new tests,
zero mypy errors, ruff clean.

Tracks:
- transitions: add add_gradient_wipe / add_card_swipe / add_cube_rotate / add_swap
- visual_effects: register Sepia, Vignette, Reflection, StaticNoise, Tiling,
  TornEdge, CRTMonitor, Mosaic, OutlineEdges, WindowSpotlight, ColorTint,
  Border, Colorize as typed @register_effect classes
- cursor_effects: 33 new cursor Effect classes (CursorHighlight,
  CursorSpotlight, CursorMagnify, CursorGlow, CursorColor, CursorIsolation,
  CursorGradient, CursorLens, CursorNegative, CursorSmoothing, plus
  LeftClickBurst1-4/RightClickBurst1-4, Left/RightClickZoom/Rings/Ripple/
  Scope/Target/Warp/Sound and RightClickScaling)
- audio_effects: AudioCompression, Pitch (Mac-only), ClipSpeedAudio;
  extend NoiseRemoval with sensitivity/reduction params
- emphasize_builder: BaseClip.add_emphasize(ramp_position, ramp_in_seconds,
  ramp_out_seconds, intensity) wrapping existing Emphasize class
- audiate_language: writable AudiateProject.language setter
- markers_api: MarkerList.rename / move / remove_by_name / next_after /
  prev_before
- media_bin_api: Media.rename; MediaBin.unused_media / delete_unused /
  sorted / import_folder / import_many
- track_api: Track.add_placeholder / freeze_at_clip_start / freeze_at_clip_end
  / extend_clip_to; PlaceholderMedia.title property
- watermark_api: Project.add_watermark extended with scale/x_offset/y_offset;
  new Project.add_text_watermark; builder forwards all new params

Fixes: mypy Effect.duration return type widened to float|None for
cursor-effect subclass compatibility; video_production.py type narrowing;
recording_sync.py explicit dict[str, Any] return.

Verified: ruff clean, mypy 0 errors in 77 files, 3607 tests pass (up from 3322).

+++ src/camtasia/effects/base.py
@@ -103,7 +103,7 @@ class Effect:
         return self._data.get('start')
 
     @property
-    def duration(self) -> int | None:
+    def duration(self) -> float | None:
         """Effect duration in ticks, or ``None`` if not time-bounded."""
         return self._data.get('duration')
